Backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-dotenv==1.0.0
httpx==0.25.1
pydantic==2.11.4
langchain==0.3.25
langchain-core==0.3.59
langchain-google-genai==2.1.4
google-generativeai==0.3.1